Remove unnecessary assertions
No longer necessary because after pull request #10, the method that
gets called subsequently (i.e. str()) does not panic, so there is no
need for additional steps for preventing a panic.

Also, the assertions were ineffective in preventing a panic in the old
code: ensure_type(UTF8) succeeds for Utf8 and for Utf8Bytes, but utf8()
(which used to be called before PR #10) would panic upon Utf8Bytes.